requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
geopy==2.3.0